FIT-22: Move files around to be more clean
